externalize examples
