+++ examples/asynchronous.rs
