+++ examples/synchronous.rs
