+++ fastrace-opentelemetry/Cargo.toml
@@ -16,13 +16,9 @@ rust-version.workspace = true
 [dependencies]
 fastrace = { workspace = true }
 log = { workspace = true }
-opentelemetry = { version = "0.29.0", default-features = false, features = [
-  "trace",
-] }
-opentelemetry_sdk = { version = "0.29.0", default-features = false, features = [
-  "trace",
-] }
+opentelemetry = { workspace = true }
+opentelemetry_sdk = { workspace = true }
 pollster = { version = "0.4.0" }
 
 [dev-dependencies]
-opentelemetry-otlp = { version = "0.29.0", features = ["trace"] }
+opentelemetry-otlp = { workspace = true }
