+++ fastrace/Cargo.toml
@@ -30,29 +30,17 @@ async-trait = "0.1"
 crossbeam = "0.8"
 divan = "0.1"
 fastrace = { path = ".", features = ["enable"] }
-fastrace-datadog = { workspace = true }
-fastrace-jaeger = { workspace = true }
 fastrace-opentelemetry = { workspace = true }
 flume = "0.11"
 futures-timer = "3"
-log = { workspace = true }
-logcall = "0.1"
-logforth = { version = "0.23.1", features = ["fastrace"] }
 mockall = "0.13"
-opentelemetry = { version = "0.29.0", features = ["trace"] }
-opentelemetry-otlp = { version = "0.29.0", features = ["trace", "grpc-tonic"] }
-opentelemetry_sdk = { version = "0.29.0", features = ["trace"] }
 pollster = { version = "0.4.0" }
 rand = "0.9"
-rustracing = "0.6"
 serial_test = "3.1"
-test-harness = "0.3"
-tokio = { version = "1.38", features = [
-  "rt",
-  "time",
-  "macros",
-  "rt-multi-thread",
-] }
+opentelemetry = { workspace = true }
+opentelemetry-otlp = { workspace = true }
+opentelemetry_sdk = { workspace = true }
+tokio = { workspace = true }
 tracing = "0.1"
 tracing-opentelemetry = { version = "0.30.0" }
 tracing-subscriber = { version = "0.3.19" }
